services/api/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
supabase>=2.10.0
PyJWT[crypto]>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
